# server/openapiv2/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "group.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/balance/api/delete/group": {
      "post": {
        "operationId": "GroupService_DeleteGroup",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoDeleteGroupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoDeleteGroupRequest"
            }
          }
        ],
        "tags": [
          "GroupService"
        ]
      }
    },
...
    "/balance/api/get/group": {
      "post": {
        "operationId": "GroupService_GetGroup",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoGetGroupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoGetGroupRequest"
            }
          }
        ],
        "tags": [
          "GroupService"
        ]
      }
    },
...
    "/balance/api/list/group": {
      "post": {
        "operationId": "GroupService_ListGroup",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoListGroupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoListGroupRequest"
            }
          }
        ],
        "tags": [
          "GroupService"
        ]
      }
    },
...
    "/balance/api/register/group": {
      "post": {
        "operationId": "GroupService_RegisterGroup",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoRegisterGroupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoRegisterGroupRequest"
            }
          }
        ],
        "tags": [
          "GroupService"
        ]
      }
    },
...
    "/balance/api/update/group": {
      "post": {
        "operationId": "GroupService_UpdateGroup",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoUpdateGroupResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoUpdateGroupRequest"
            }
          }
        ],
        "tags": [
          "GroupService"
        ]
      }
    },
...
  },
  "definitions": {
    "protoDeleteGroupRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoDeleteGroupResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        }
      }
    },
...
    "protoGetGroupRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoGetGroupResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        }
      }
    },
...
    "protoGroupResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        }
      }
    },
...
    "protoListGroupRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoListGroupResponse": {
      "type": "object",
      "properties": {
        "groupList": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protoGroupResponse"
          }
        }
      }
    },
...
    "protoRegisterGroupRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoRegisterGroupResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        }
      }
    },
...
    "protoUpdateGroupRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        },
        "sessoinId": {
          "type": "string"
        }
      }
    },
    "protoUpdateGroupResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string"
        },
        "groupName": {
          "type": "string"
        }
      }
    },
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}